Refactor overall structure.
