+++ src/main/java/com/github/dakusui/actionunit/actions/cmd/CommanderOption.java
@@ -10,12 +10,12 @@ public interface CommanderOption {
   }
 
   @SuppressWarnings("unchecked")
-  default <B extends Commander> B addTo(B builder, String value, boolean longFormat) {
+  default <B extends CompatCommander> B addTo(B builder, String value, boolean longFormat) {
     return (B) builder.add(String.format("%s %s", this.format(longFormat), CommanderUtils.quoteWithSingleQuotesForShell(value)));
   }
 
   @SuppressWarnings("unchecked")
-  default <B extends Commander> B addTo(B builder, boolean longFormat) {
+  default <B extends CompatCommander> B addTo(B builder, boolean longFormat) {
     return (B) builder.add(this.format(longFormat));
   }
 }
