+++ src/main/java/com/github/dakusui/actionunit/core/context/ContextFunction.java
@@ -3,14 +3,9 @@ package com.github.dakusui.actionunit.core.context;
 import com.github.dakusui.actionunit.core.Context;
 import com.github.dakusui.printables.PrintableFunction;
 
-import java.util.function.BiFunction;
 import java.util.function.Function;
-import java.util.function.IntFunction;
 import java.util.function.Supplier;
 
-import static com.github.dakusui.actionunit.core.context.ContextFunctions.PLACE_HOLDER_FORMATTER;
-import static com.github.dakusui.actionunit.core.context.ContextFunctions.describeFunctionalObject;
-import static com.github.dakusui.printables.Printables.function;
 import static java.util.Objects.requireNonNull;
 
 @FunctionalInterface
@@ -24,15 +19,6 @@ public interface ContextFunction<R> extends Function<Context, R>, Printable {
     return (Context r) -> after.apply(apply(r));
   }
 
-
-  static <T, R> ContextFunction<R> of(String variableName, Function<T, R> function) {
-    return ContextFunctions.<R>contextFunctionFor(variableName)
-        .with(function(
-            (Params params) -> function.apply(params.valueOf(variableName))
-        ).describe(
-            function.toString()
-        ));
-  }
 
   static <R> ContextFunction<R> of(Supplier<String> formatter, Function<Context, R> function) {
     return new ContextFunction.Impl<R>(formatter, function);
@@ -65,35 +51,4 @@ public interface ContextFunction<R> extends Function<Context, R>, Printable {
     }
   }
 
-  class Builder<R> {
-    private final String[]                               variableNames;
-    private final BiFunction<Function, String[], String> descriptionFormatter;
-
-    public Builder(String... variableNames) {
-      this(
-          (f, v) -> describeFunctionalObject(f, PLACE_HOLDER_FORMATTER, v),
-          variableNames
-      );
-    }
-
-    public Builder(
-        BiFunction<Function, String[], String> descriptionFormatter,
-        String... variableNames) {
-      this.descriptionFormatter = requireNonNull(descriptionFormatter);
-      this.variableNames = requireNonNull(variableNames);
-    }
-
-    public Builder(
-        IntFunction<String> placeHolderFormatter,
-        String... variableNames) {
-      this((f, v) -> describeFunctionalObject(f, placeHolderFormatter, v), variableNames);
-    }
-
-    public ContextFunction<R> with(Function<Params, R> function) {
-      requireNonNull(function);
-      return new Impl<>(
-          () -> descriptionFormatter.apply(function, variableNames),
-          (Context c) -> function.apply(Params.create(c, variableNames)));
-    }
-  }
 }
