+++ src/main/java/com/github/dakusui/actionunit/core/context/ContextFunctions.java
@@ -1,5 +1,8 @@
 package com.github.dakusui.actionunit.core.context;
 
+import com.github.dakusui.actionunit.core.context.multiparams.MultiParamsConsumerBuilder;
+import com.github.dakusui.actionunit.core.context.multiparams.MultiParamsFunctionBuilder;
+import com.github.dakusui.actionunit.core.context.multiparams.MultiParamsPredicateBuilder;
 import com.github.dakusui.actionunit.utils.StableTemplatingUtils;
 
 import java.io.PrintStream;
@@ -19,19 +22,19 @@ public enum ContextFunctions {
 
   public static final IntFunction<String> PLACE_HOLDER_FORMATTER = i -> String.format("{{%s}}", i);
 
-  public static ContextPredicate.Builder contextPredicateFor(String... variableNames) {
-    return new ContextPredicate.Builder(variableNames);
+  public static MultiParamsPredicateBuilder contextPredicateFor(String... variableNames) {
+    return new MultiParamsPredicateBuilder(variableNames);
   }
 
-  public static ContextConsumer.Builder contextConsumerFor(String... variableNames) {
-    return new ContextConsumer.Builder(variableNames);
+  public static MultiParamsConsumerBuilder contextConsumerFor(String... variableNames) {
+    return new MultiParamsConsumerBuilder(variableNames);
   }
 
-  public static <R> ContextFunction.Builder<R> contextFunctionFor(String... variableNames) {
-    return new ContextFunction.Builder<>(variableNames);
+  public static <R> MultiParamsFunctionBuilder<R> contextFunctionFor(String... variableNames) {
+    return new MultiParamsFunctionBuilder<>(variableNames);
   }
 
-  static String describeFunctionalObject(Object f, final IntFunction<String> placeHolderFormatter, String... v) {
+  public static String describeFunctionalObject(Object f, final IntFunction<String> placeHolderFormatter, String... v) {
     return String.format("(%s)->%s",
         String.join(",", v),
         StableTemplatingUtils.template(
