+++ src/main/java/com/github/dakusui/actionunit/core/context/multiparams/Params.java
@@ -1,4 +1,4 @@
-package com.github.dakusui.actionunit.core.context;
+package com.github.dakusui.actionunit.core.context.multiparams;
 
 import com.github.dakusui.actionunit.core.Context;
