+++ src/main/java/com/github/dakusui/actionunit/linux/Cat.java
@@ -1,7 +1,8 @@
 package com.github.dakusui.actionunit.linux;
 
-import com.github.dakusui.actionunit.core.Context;
+import com.github.dakusui.actionunit.actions.cmd.CommandLineComposer;
 import com.github.dakusui.actionunit.actions.cmd.Commander;
+import com.github.dakusui.actionunit.core.Context;
 
 import java.io.File;
 
@@ -13,19 +14,36 @@ public class Cat extends Commander<Cat> {
   }
 
   public Cat number() {
-    return this.add("-n");
+//    return this.add("-n");
+    return this;
   }
 
   public Cat file(String fileName) {
-    return this.add(requireNonNull(fileName));
+    //return this.add(requireNonNull(fileName));
+    return this;
   }
 
   public Cat file(File file) {
-    return this.add(requireNonNull(file).getAbsolutePath());
+    //return this.add(requireNonNull(file).getAbsolutePath());
+    return this;
+  }
+
+  public Cat file() {
+    return this;
   }
 
   @Override
-  protected String program() {
-    return "/bin/cat";
+  protected CommandLineComposer commandLineComposer() {
+    return new CommandLineComposer() {
+      @Override
+      public String commandLineString() {
+        return "/bin/cat";
+      }
+    };
+  }
+
+  @Override
+  protected String[] variableNames() {
+    return new String[0];
   }
 }
