+++ src/test/java/com/github/dakusui/actionunit/extras/cmd/CompatCommanderSupportTest.java
@@ -8,12 +8,12 @@ import org.junit.Test;
 
 import java.util.stream.Stream;
 
-import static com.github.dakusui.actionunit.actions.cmd.Commander.commander;
+import static com.github.dakusui.actionunit.actions.cmd.CompatCommander.commander;
 import static com.github.dakusui.actionunit.core.ActionSupport.cmd;
 import static com.github.dakusui.actionunit.core.ActionSupport.forEach;
 import static com.github.dakusui.actionunit.core.ActionSupport.leaf;
 
-public class CommanderSupportTest extends TestUtils.TestBase {
+public class CompatCommanderSupportTest extends TestUtils.TestBase {
 
   @Test
   public void test() {
